+++ cmd/faceblur/main.go
@@ -34,7 +34,7 @@ func main() {
 	xmlFile := os.Args[2]
 
 	// open webcam
-	webcam, err := gocv.VideoCaptureDevice(int(deviceID))
+	webcam, err := gocv.VideoCaptureDevice(deviceID)
 	if err != nil {
 		fmt.Printf("error opening video capture device: %v\n", deviceID)
 		return