+++ cmd/mjpeg-streamer/main.go
@@ -62,7 +62,7 @@ func main() {
 	host := os.Args[2]
 
 	// open webcam
-	webcam, err = gocv.VideoCaptureDevice(int(deviceID))
+	webcam, err = gocv.VideoCaptureDevice(deviceID)
 	if err != nil {
 		fmt.Printf("error opening video capture device: %v\n", deviceID)
 		return