+++ core.h
@@ -36,6 +36,21 @@ typedef struct Point {
   int y;
 } Point;
 
+// Wrapper for the vector of Point structs aka std::vector<Point>
+typedef struct Points {
+  Point* points;
+  int length;
+} Points;
+
+// Contour is alias for Points
+typedef Points Contour;
+
+// Wrapper for the vector of Points vectors aka std::vector< std::vector<Point> >
+typedef struct Contours {
+  Contour* contours;
+  int length;
+} Contours;
+
 // Wrapper for an individual cv::cvRect
 typedef struct Rect {
   int x;
@@ -63,16 +78,49 @@ typedef struct Scalar {
   double val3;
   double val4;
 } Scalar;
-  
+
+// Wrapper for an individual cv::Moment
+typedef struct Moment {
+  double m00;
+  double m10;
+  double m01;
+  double m20;
+  double m11;
+  double m02;
+  double m30;
+  double m21;
+  double m12;
+  double m03;
+
+  double mu20;
+  double mu11;
+  double mu02;
+  double mu30;
+  double mu21;
+  double mu12;
+  double mu03;
+
+  double nu20;
+  double nu11;
+  double nu02;
+  double nu30;
+  double nu21;
+  double nu12;
+  double nu03;
+} Moment;
+
 #ifdef __cplusplus
 typedef cv::Mat* Mat;
+typedef cv::TermCriteria* TermCriteria;
 #else
 typedef void* Mat;
+typedef void* TermCriteria;
 #endif
   
 struct ByteArray toByteArray(const char* buf, int len);
 void ByteArray_Release(struct ByteArray buf);
 
+void Contours_Close(struct Contours cs);
 void Rects_Close(struct Rects rs);
 
 Mat Mat_New();
@@ -105,6 +153,8 @@ int Mat_GetOptimalDFTSize(int vecsize);
 void Mat_DFT(Mat m, Mat dst);
 void Mat_Merge(Mat m, size_t count, Mat dst);
 void Mat_Normalize(Mat src, Mat dst, double alpha, double beta, int typ);
+
+TermCriteria TermCriteria_New(int typ, int maxCount, double epsilon);
 
 #ifdef __cplusplus
 }