+++ imgproc.go
@@ -8,6 +8,7 @@ import "C"
 import (
 	"image"
 	"image/color"
+	"reflect"
 	"unsafe"
 )
 
@@ -67,6 +68,183 @@ func Dilate(src Mat, dst Mat, kernel Mat) {
 //
 func Erode(src Mat, dst Mat, kernel Mat) {
 	C.Erode(src.p, dst.p, kernel.p)
+}
+
+// RetrievalMode is the mode of the contour retrieval algorithm.
+type RetrievalMode int
+
+const (
+	// RetrievalExternal retrieves only the extreme outer contours.
+	// It sets `hierarchy[i][2]=hierarchy[i][3]=-1` for all the contours.
+	RetrievalExternal RetrievalMode = 0
+
+	// RetrievalList retrieves all of the contours without establishing
+	// any hierarchical relationships.
+	RetrievalList = 1
+
+	// RetrievalCComp retrieves all of the contours and organizes them into
+	// a two-level hierarchy. At the top level, there are external boundaries
+	// of the components. At the second level, there are boundaries of the holes.
+	// If there is another contour inside a hole of a connected component, it
+	// is still put at the top level.
+	RetrievalCComp = 2
+
+	// RetrievalTree retrieves all of the contours and reconstructs a full
+	// hierarchy of nested contours.
+	RetrievalTree = 3
+
+	// RetrievalFloodfill lacks a description in the original header.
+	RetrievalFloodfill = 4
+)
+
+// ContourApproximationMode is the mode of the contour approximation algorithm.
+type ContourApproximationMode int
+
+const (
+	// ChainApproxNone stores absolutely all the contour points. That is,
+	// any 2 subsequent points (x1,y1) and (x2,y2) of the contour will be
+	// either horizontal, vertical or diagonal neighbors, that is,
+	// max(abs(x1-x2),abs(y2-y1))==1.
+	ChainApproxNone ContourApproximationMode = 1
+
+	// ChainApproxSimple compresses horizontal, vertical, and diagonal segments
+	// and leaves only their end points.
+	// For example, an up-right rectangular contour is encoded with 4 points.
+	ChainApproxSimple = 2
+
+	// ChainApproxTC89L1 applies one of the flavors of the Teh-Chin chain
+	// approximation algorithms.
+	ChainApproxTC89L1 = 3
+
+	// ChainApproxTC89KCOS applies one of the flavors of the Teh-Chin chain
+	// approximation algorithms.
+	ChainApproxTC89KCOS = 4
+)
+
+// BoundingRect calculates the up-right bounding rectangle of a point set.
+//
+// For further details, please see:
+// https://docs.opencv.org/3.3.0/d3/dc0/group__imgproc__shape.html#gacb413ddce8e48ff3ca61ed7cf626a366
+//
+func BoundingRect(contour []image.Point) image.Rectangle {
+	cPointArray := make([]C.struct_Point, len(contour))
+	for i, r := range contour {
+		cPoint := C.struct_Point{
+			x: C.int(r.X),
+			y: C.int(r.Y),
+		}
+		cPointArray[i] = cPoint
+	}
+
+	cContour := C.struct_Points{
+		points: (*C.Point)(&cPointArray[0]),
+		length: C.int(len(contour)),
+	}
+
+	r := C.BoundingRect(cContour)
+	rect := image.Rect(int(r.x), int(r.y), int(r.x+r.width), int(r.y+r.height))
+	return rect
+}
+
+// ContourArea calculates a contour area.
+//
+// For further details, please see:
+// https://docs.opencv.org/3.3.0/d3/dc0/group__imgproc__shape.html#ga2c759ed9f497d4a618048a2f56dc97f1
+//
+func ContourArea(contour []image.Point) float64 {
+	cPointArray := make([]C.struct_Point, len(contour))
+	for i, r := range contour {
+		cPoint := C.struct_Point{
+			x: C.int(r.X),
+			y: C.int(r.Y),
+		}
+		cPointArray[i] = cPoint
+	}
+
+	cContour := C.struct_Points{
+		points: (*C.Point)(&cPointArray[0]),
+		length: C.int(len(contour)),
+	}
+
+	result := C.ContourArea(cContour)
+	return float64(result)
+}
+
+// FindContours finds contours in a binary image.
+//
+// For further details, please see:
+// https://docs.opencv.org/3.3.0/d3/dc0/group__imgproc__shape.html#ga17ed9f5d79ae97bd4c7cf18403e1689a
+//
+func FindContours(src Mat, mode RetrievalMode, method ContourApproximationMode) [][]image.Point {
+	ret := C.FindContours(src.p, C.int(mode), C.int(method))
+	defer C.Contours_Close(ret)
+
+	cArray := ret.contours
+	cLength := int(ret.length)
+	cHdr := reflect.SliceHeader{
+		Data: uintptr(unsafe.Pointer(cArray)),
+		Len:  cLength,
+		Cap:  cLength,
+	}
+	sContours := *(*[]C.Points)(unsafe.Pointer(&cHdr))
+
+	contours := make([][]image.Point, cLength)
+	for i, pts := range sContours {
+		pArray := pts.points
+		pLength := int(pts.length)
+		pHdr := reflect.SliceHeader{
+			Data: uintptr(unsafe.Pointer(pArray)),
+			Len:  pLength,
+			Cap:  pLength,
+		}
+		sPoints := *(*[]C.Point)(unsafe.Pointer(&pHdr))
+
+		points := make([]image.Point, pLength)
+		for j, pt := range sPoints {
+			points[j] = image.Pt(int(pt.x), int(pt.y))
+		}
+		contours[i] = points
+	}
+
+	return contours
+}
+
+// Moments calculates all of the moments up to the third order of a polygon
+// or rasterized shape.
+//
+// For further details, please see:
+// https://docs.opencv.org/3.3.1/d3/dc0/group__imgproc__shape.html#ga556a180f43cab22649c23ada36a8a139
+//
+func Moments(src Mat, binaryImage bool) map[string]float64 {
+	r := C.Moments(src.p, C.bool(binaryImage))
+
+	result := make(map[string]float64)
+	result["m00"] = float64(r.m00)
+	result["m10"] = float64(r.m10)
+	result["m01"] = float64(r.m01)
+	result["m20"] = float64(r.m20)
+	result["m11"] = float64(r.m11)
+	result["m02"] = float64(r.m02)
+	result["m30"] = float64(r.m30)
+	result["m21"] = float64(r.m21)
+	result["m12"] = float64(r.m12)
+	result["m03"] = float64(r.m03)
+	result["mu20"] = float64(r.mu20)
+	result["mu11"] = float64(r.mu11)
+	result["mu02"] = float64(r.mu02)
+	result["mu30"] = float64(r.mu30)
+	result["mu21"] = float64(r.mu21)
+	result["mu12"] = float64(r.mu12)
+	result["mu03"] = float64(r.mu03)
+	result["nu20"] = float64(r.nu20)
+	result["nu11"] = float64(r.nu11)
+	result["nu02"] = float64(r.nu02)
+	result["nu30"] = float64(r.nu30)
+	result["nu21"] = float64(r.nu21)
+	result["nu12"] = float64(r.nu12)
+	result["nu03"] = float64(r.nu03)
+
+	return result
 }
 
 // MorphologyEx performs advanced morphological transformations.
@@ -136,6 +314,32 @@ const (
 	MorphHitmiss = 7
 )
 
+// BorderType type of border.
+type BorderType int
+
+const (
+	// BorderConstant border type
+	BorderConstant BorderType = 0
+
+	// BorderReplicate border type
+	BorderReplicate = 1
+
+	// BorderReflect border type
+	BorderReflect = 2
+
+	// BorderWrap border type
+	BorderWrap = 3
+
+	// BorderReflect101 border type
+	BorderReflect101 = 4
+
+	// BorderTransparent border type
+	BorderTransparent = 5
+
+	// BorderDefault border type
+	BorderDefault = BorderReflect101
+)
+
 // GaussianBlur blurs an image Mat using a Gaussian filter.
 // The function convolves the src Mat image into the dst Mat using
 // the specified Gaussian kernel params.
@@ -144,13 +348,33 @@ const (
 // http://docs.opencv.org/3.3.1/d4/d86/group__imgproc__filter.html#gaabe8c836e97159a9193fb0b11ac52cf1
 //
 func GaussianBlur(src Mat, dst Mat, ksize image.Point, sigmaX float64,
-	sigmaY float64, borderType int) {
+	sigmaY float64, borderType BorderType) {
 	pSize := C.struct_Size{
 		height: C.int(ksize.X),
 		width:  C.int(ksize.Y),
 	}
 
 	C.GaussianBlur(src.p, dst.p, pSize, C.double(sigmaX), C.double(sigmaY), C.int(borderType))
+}
+
+// Laplacian calculates the Laplacian of an image.
+//
+// For further details, please see:
+// https://docs.opencv.org/3.3.1/d4/d86/group__imgproc__filter.html#gad78703e4c8fe703d479c1860d76429e6
+//
+func Laplacian(src Mat, dst Mat, dDepth int, kSize int, scale float64,
+	delta float64, borderType BorderType) {
+	C.Laplacian(src.p, dst.p, C.int(dDepth), C.int(kSize), C.double(scale), C.double(delta), C.int(borderType))
+}
+
+// Scharr calculates the first x- or y- image derivative using Scharr operator.
+//
+// For further details, please see:
+// https://docs.opencv.org/3.3.1/d4/d86/group__imgproc__filter.html#gaa13106761eedf14798f37aa2d60404c9
+//
+func Scharr(src Mat, dst Mat, dDepth int, dx int, dy int, scale float64,
+	delta float64, borderType BorderType) {
+	C.Scharr(src.p, dst.p, C.int(dDepth), C.int(dx), C.int(dy), C.double(scale), C.double(delta), C.int(borderType))
 }
 
 // MedianBlur blurs an image using the median filter.
@@ -175,6 +399,37 @@ func MedianBlur(src Mat, dst Mat, ksize int) {
 //
 func Canny(src Mat, edges Mat, t1 float32, t2 float32) {
 	C.Canny(src.p, edges.p, C.double(t1), C.double(t2))
+}
+
+// CornerSubPix Refines the corner locations. The function iterates to find
+// the sub-pixel accurate location of corners or radial saddle points.
+//
+// For further details, please see:
+// https://docs.opencv.org/3.3.1/dd/d1a/group__imgproc__feature.html#ga354e0d7c86d0d9da75de9b9701a9a87e
+//
+func CornerSubPix(img Mat, corners Mat, winSize image.Point, zeroZone image.Point, criteria TermCriteria) {
+	winSz := C.struct_Size{
+		height: C.int(winSize.X),
+		width:  C.int(winSize.Y),
+	}
+
+	zeroSz := C.struct_Size{
+		height: C.int(zeroZone.X),
+		width:  C.int(zeroZone.Y),
+	}
+
+	C.CornerSubPix(img.p, corners.p, winSz, zeroSz, criteria.p)
+	return
+}
+
+// GoodFeaturesToTrack determines strong corners on an image. The function
+// finds the most prominent corners in the image or in the specified image region.
+//
+// For further details, please see:
+// https://docs.opencv.org/3.3.1/dd/d1a/group__imgproc__feature.html#ga1d6bb77486c8f92d79c8793ad995d541
+//
+func GoodFeaturesToTrack(img Mat, corners Mat, maxCorners int, quality float64, minDist float64) {
+	C.GoodFeaturesToTrack(img.p, corners.p, C.int(maxCorners), C.double(quality), C.double(minDist))
 }
 
 // HoughCircles finds circles in a grayscale image using the Hough transform.
@@ -207,6 +462,44 @@ func HoughLines(src Mat, lines Mat, rho float32, theta float32, threshold int) {
 //
 func HoughLinesP(src Mat, lines Mat, rho float32, theta float32, threshold int) {
 	C.HoughLinesP(src.p, lines.p, C.double(rho), C.double(theta), C.int(threshold))
+}
+
+// ThresholdType type of threshold operation.
+type ThresholdType int
+
+const (
+	// ThresholdBinary threshold type
+	ThresholdBinary ThresholdType = 0
+
+	// ThresholdBinaryInv threshold type
+	ThresholdBinaryInv = 1
+
+	// ThresholdTrunc threshold type
+	ThresholdTrunc = 2
+
+	// ThresholdToZero threshold type
+	ThresholdToZero = 3
+
+	// ThresholdToZeroInv threshold type
+	ThresholdToZeroInv = 4
+
+	// ThresholdMask threshold type
+	ThresholdMask = 7
+
+	// ThresholdOtsu threshold type
+	ThresholdOtsu = 8
+
+	// ThresholdTriangle threshold type
+	ThresholdTriangle = 16
+)
+
+// Threshold applies a fixed-level threshold to each array element.
+//
+// For further details, please see:
+// https://docs.opencv.org/3.3.0/d7/d1b/group__imgproc__misc.html#gae8a4a146d1ca78c626a53577199e9c57
+//
+func Threshold(src Mat, dst Mat, thresh float32, maxvalue float32, typ ThresholdType) {
+	C.Threshold(src.p, dst.p, C.double(thresh), C.double(maxvalue), C.int(typ))
 }
 
 // ArrowedLine draws a arrow segment pointing from the first point