+++ objdetect.cpp
@@ -16,7 +16,24 @@ int CascadeClassifier_Load(CascadeClassifier cs, const char* name) {
   
 struct Rects CascadeClassifier_DetectMultiScale(CascadeClassifier cs, Mat img) {
     std::vector<cv::Rect> detected;
-    cs->detectMultiScale(*img, detected); // TODO control default parameter
+    cs->detectMultiScale(*img, detected); // uses all default parameters
+    Rect* rects = new Rect[detected.size()];
+    for (size_t i = 0; i < detected.size(); ++i) {
+      Rect r = {detected[i].x, detected[i].y, detected[i].width, detected[i].height};
+      rects[i] = r;
+    }
+    Rects ret = {rects, (int)detected.size()};
+    return ret;
+}
+
+struct Rects CascadeClassifier_DetectMultiScaleWithParams(CascadeClassifier cs, Mat img,
+    double scale, int minNeighbors, int flags, Size minSize, Size maxSize) {
+
+    cv::Size minSz(minSize.width, minSize.height);
+    cv::Size maxSz(maxSize.width, maxSize.height);
+
+    std::vector<cv::Rect> detected;
+    cs->detectMultiScale(*img, detected, scale, minNeighbors, flags, minSz, maxSz);
     Rect* rects = new Rect[detected.size()];
     for (size_t i = 0; i < detected.size(); ++i) {
       Rect r = {detected[i].x, detected[i].y, detected[i].width, detected[i].height};
@@ -42,7 +59,25 @@ int HOGDescriptor_Load(HOGDescriptor hog, const char* name) {
 
 struct Rects HOGDescriptor_DetectMultiScale(HOGDescriptor hog, Mat img) {
     std::vector<cv::Rect> detected;
-    hog->detectMultiScale(*img, detected); // TODO control default parameter
+    hog->detectMultiScale(*img, detected);
+    Rect* rects = new Rect[detected.size()];
+    for (size_t i = 0; i < detected.size(); ++i) {
+      Rect r = {detected[i].x, detected[i].y, detected[i].width, detected[i].height};
+      rects[i] = r;
+    }
+    Rects ret = {rects, (int)detected.size()};
+    return ret;
+}
+
+struct Rects HOGDescriptor_DetectMultiScaleWithParams(HOGDescriptor hog, Mat img,
+    double hitThresh, Size winStride, Size padding, double scale, double finalThresh, 
+    bool useMeanshiftGrouping) {
+
+    cv::Size wSz(winStride.width, winStride.height);
+    cv::Size pSz(padding.width, padding.height);
+
+    std::vector<cv::Rect> detected;
+    hog->detectMultiScale(*img, detected, hitThresh, wSz, pSz, scale, finalThresh, useMeanshiftGrouping);
     Rect* rects = new Rect[detected.size()];
     for (size_t i = 0; i < detected.size(); ++i) {
       Rect r = {detected[i].x, detected[i].y, detected[i].width, detected[i].height};