+++ CHANGELOG.md
@@ -1,3 +1,54 @@
+0.5.0
+---
+* **core**
+    * add TermCriteria for iterative algorithms
+* **imgproc**
+    * add CornerSubPix() and GoodFeaturesToTrack() for corner detection
+* **objdetect**
+    * add DetectMultiScaleWithParams() for HOGDescriptor
+    * add DetectMultiScaleWithParams() to allow override of defaults for CascadeClassifier
+* **video**
+    * add CalcOpticalFlowFarneback() for Farneback optical flow calculations
+    * add CalcOpticalFlowPyrLK() for Lucas-Kanade optical flow calculations
+* **videoio**
+    * use temp directory for Windows test compat.
+* **build**
+    * enable Appveyor build w/cache
+* **osx**
+    * update env path to always match installed OpenCV from Homebrew
+
+0.4.0
+---
+* **core**
+    * Added cv::mean binding with single argument
+    * fix the write-strings warning
+    * return temp pointer fix
+* **examples**
+    * add counter example
+    * add motion-detect command
+    * correct counter
+    * remove redundant cast and other small cleanup
+    * set motion detect example to fullscreen
+    * use MOG2 for continous motion detection, instead of simplistic first frame only
+* **highgui**
+    * ability to better control the fullscreen window
+* **imgproc**
+    * add BorderType param type for GaussianBlur
+    * add BoundingRect() function
+    * add ContourArea() function
+    * add FindContours() function along with associated data types
+    * add Laplacian and Scharr functions
+    * add Moments() function
+    * add Threshold function
+* **pvl**
+    * add needed lib for linker missing in README
+* **test**
+    * slightly more permissive version test
+* **videoio**
+    * Add image compression flags for gocv.IMWrite
+    * Fixed possible looping out of compression parameters length
+    * Make dedicated function to run cv::imwrite with compression parameters
+
 0.3.1
 ---
 * **overall**